requirements/testing.txt
pytest==3.3.1
pytest-django==3.1.2
mock==2.0.0